fixes for new replication features

+++ arangod/RestHandler/RestReplicationHandler.cpp
@@ -3911,9 +3911,9 @@ void RestReplicationHandler::handleCommandMakeSlave () {
   config._verbose                 = JsonHelper::getBooleanValue(json.get(), "verbose", defaults._verbose);
   config._requireFromPresent      = JsonHelper::getBooleanValue(json.get(), "requireFromPresent", defaults._requireFromPresent);
   config._restrictType            = JsonHelper::getStringValue(json.get(), "restrictType", defaults._restrictType);
-  config._connectionRetryWaitTime = static_cast<uint64_t>(JsonHelper::getNumericValue<double>(json.get(), "connectionRetryWaitTime", defaults._connectionRetryWaitTime / (1000 * 1000)));
-  config._idleMinWaitTime         = static_cast<uint64_t>(JsonHelper::getNumericValue<double>(json.get(), "idleMinWaitTime", defaults._idleMinWaitTime / (1000 * 1000)));
-  config._idleMaxWaitTime         = static_cast<uint64_t>(JsonHelper::getNumericValue<double>(json.get(), "idleMaxWaitTime", defaults._idleMaxWaitTime / (1000 * 1000)));
+  config._connectionRetryWaitTime = static_cast<uint64_t>(JsonHelper::getNumericValue<double>(json.get(), "connectionRetryWaitTime", static_cast<double>(defaults._connectionRetryWaitTime) / (1000.0 * 1000.0)));
+  config._idleMinWaitTime         = static_cast<uint64_t>(JsonHelper::getNumericValue<double>(json.get(), "idleMinWaitTime", static_cast<double>(defaults._idleMinWaitTime) / (1000.0 * 1000.0)));
+  config._idleMaxWaitTime         = static_cast<uint64_t>(JsonHelper::getNumericValue<double>(json.get(), "idleMaxWaitTime", static_cast<double>(defaults._idleMaxWaitTime) / (1000.0 * 1000.0)));
   
   TRI_json_t* restriction = JsonHelper::getObjectElement(json.get(), "restrictCollections");
 
@@ -4611,9 +4611,9 @@ void RestReplicationHandler::handleCommandApplierSetConfig () {
   config._verbose                 = JsonHelper::getBooleanValue(json.get(), "verbose", config._verbose);
   config._requireFromPresent      = JsonHelper::getBooleanValue(json.get(), "requireFromPresent", config._requireFromPresent);
   config._restrictType            = JsonHelper::getStringValue(json.get(), "restrictType", config._restrictType);
-  config._connectionRetryWaitTime = static_cast<uint64_t>(JsonHelper::getNumericValue<double>(json.get(), "connectionRetryWaitTime", config._connectionRetryWaitTime / (1000 * 1000)));
-  config._idleMinWaitTime         = static_cast<uint64_t>(JsonHelper::getNumericValue<double>(json.get(), "idleMinWaitTime", config._idleMinWaitTime / (1000 * 1000)));
-  config._idleMaxWaitTime         = static_cast<uint64_t>(JsonHelper::getNumericValue<double>(json.get(), "idleMaxWaitTime", config._idleMaxWaitTime / (1000 * 1000)));
+  config._connectionRetryWaitTime = static_cast<uint64_t>(JsonHelper::getNumericValue<double>(json.get(), "connectionRetryWaitTime", static_cast<double>(config._connectionRetryWaitTime) / (1000.0 * 1000.0)));
+  config._idleMinWaitTime         = static_cast<uint64_t>(JsonHelper::getNumericValue<double>(json.get(), "idleMinWaitTime", static_cast<double>(config._idleMinWaitTime) / (1000.0 * 1000.0)));
+  config._idleMaxWaitTime         = static_cast<uint64_t>(JsonHelper::getNumericValue<double>(json.get(), "idleMaxWaitTime", static_cast<double>(config._idleMaxWaitTime) / (1000.0 * 1000.0)));
 
   value = JsonHelper::getObjectElement(json.get(), "restrictCollections");
